Separate stats library from getStats
